requirements.txt
Django==3.2.14
django-filter==22.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
Markdown==3.4.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
django-extensions
autoflake==2.0.0
isort
black
factory-boy
pytest
pytest-django